// src/gen/refs/types_pb.ts
export namespace NeoFsV2Refs {
// ...
export interface OwnerID {
  Value: Uint8Array;
}

export class OwnerIDImpl implements OwnerID {
  Value!: Uint8Array;

  constructor(data?: Partial<OwnerID>) {
    this.Value = data?.Value ?? new Uint8Array(0);
  }

  serializeBinary(): Uint8Array {
    const writer = new BinaryWriter();

    if (this.Value.length > 0) {
      writer.writeBytes(1, this.Value);
    }
    return writer.getResultBuffer();
  }

  static deserializeBinary(data: Uint8Array): OwnerIDImpl {
    const reader = new BinaryReader(data);
    const message = new OwnerIDImpl();

    // Parse protobuf wire format
    while (reader.position < reader.buffer.length) {
      const tag = reader.readVarint();
      const fieldNumber = tag >>> 3;
      const wireType = tag & 7;

      switch (fieldNumber) {
        case 1: // Value
          if (wireType === 2) { // Length-delimited
            message.Value = reader.readBytes();
          } else {
            reader.skipField(wireType);
          }
          break;
        default:
          // Skip unknown fields
          reader.skipField(wireType);
          break;
      }
    }

    return message;
  }

  toObject(): OwnerIDData {
    return {
      Value: this.Value
    };
  }

}

// Export class alias without Impl suffix for Node.js compatibility
export class OwnerID extends OwnerIDImpl {}
// ...
export interface Version {
  Major: number;
  Minor: number;
}

export class VersionImpl implements Version {
  Major!: number;
  Minor!: number;

  constructor(data?: Partial<Version>) {
    this.Major = data?.Major ?? 0;
    this.Minor = data?.Minor ?? 0;
  }

  serializeBinary(): Uint8Array {
    const writer = new BinaryWriter();

    if (this.Major !== 0) {
      writer.writeUint32(1, this.Major);
    }
    if (this.Minor !== 0) {
      writer.writeUint32(2, this.Minor);
    }
    return writer.getResultBuffer();
  }

  static deserializeBinary(data: Uint8Array): VersionImpl {
    const reader = new BinaryReader(data);
    const message = new VersionImpl();

    // Parse protobuf wire format
    while (reader.position < reader.buffer.length) {
      const tag = reader.readVarint();
      const fieldNumber = tag >>> 3;
      const wireType = tag & 7;

      switch (fieldNumber) {
        case 1: // Major
          if (wireType === 0) { // Varint
            message.Major = reader.readUint32();
          } else if (wireType === 5) { // 32-bit
            message.Major = reader.readUint32Fixed();
          } else {
            reader.skipField(wireType);
          }
          break;
        case 2: // Minor
          if (wireType === 0) { // Varint
            message.Minor = reader.readUint32();
          } else if (wireType === 5) { // 32-bit
            message.Minor = reader.readUint32Fixed();
          } else {
            reader.skipField(wireType);
          }
          break;
        default:
          // Skip unknown fields
          reader.skipField(wireType);
          break;
      }
    }

    return message;
  }

  toObject(): VersionData {
    return {
      Major: this.Major,
      Minor: this.Minor
    };
  }

}

// Export class alias without Impl suffix for Node.js compatibility
export class Version extends VersionImpl {}
// ...
} // namespace NeoFsV2Refs